requirements.txt
asgiref==3.2.10
Django==3.1.1
django-address==0.2.5
django-environ==0.4.5
django-money==1.1
Pillow==7.2.0
py-moneyed==0.8.0
pytz==2020.1
sqlparse==0.3.1
